Now user can click edit button on cards to go to create.html to view card details. A pop-up window has been added to confirm memory card deletion.

+++ source/scripts/dataDisplay.js
@@ -103,9 +103,12 @@ function deleteListener(cardElement, id, db) {
   if (deleteBtn) {
     // console.log("Inside Here");
     deleteBtn.addEventListener("click", () => {
-      deleteMemory(id, db);
-      cardElement.remove(); // Remove from DOM
-      console.log(`Deleted memory with id: ${id}`);
+      const confirmed = window.confirm("Are you sure you want to delete this memory?");
+      if (confirmed) {
+        deleteMemory(id, db);
+        cardElement.remove(); // Remove from DOM
+        console.log(`Deleted memory with id: ${id}`);
+      }
     });
   }
 }
